Define internal printf
